Implement debugNode and configNode with routing methods; refactor actor and mailbox configurations for improved clarity and functionality
