+++ internal/graph/node_impl.go
@@ -121,43 +121,3 @@ func (r *node) ProceedOnAnyRoute(mex f.Message) error {
 
 	return nil
 }
-
-type debugNode struct {
-	node
-}
-
-// OneWayRoute adds a new possible outgoing route from the node
-func (r *debugNode) OneWayRoute(name string, destination g.Node) error {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-
-	if len(r.routes) > 0 {
-		return errors.Join(g.ErrorInvalidRouting, fmt.Errorf("debugNode node [%s] already has a route", r.Name()))
-	}
-
-	r.routes[name] = route{
-		Name:        name,
-		Destination: destination,
-	}
-
-	return nil
-}
-
-type configNode struct {
-	node
-}
-
-// OneWayRoute adds a new possible outgoing route from the node
-func (r *configNode) OneWayRoute(name string, destination g.Node) error {
-	return errors.Join(g.ErrorInvalidRouting, fmt.Errorf("cannot route [%s] from the config node [%s]", name, r.Name()))
-}
-
-// TwoWayRoute adds a new possible outgoing route from the node
-func (r *configNode) TwoWayRoute(name string, destination g.Node) error {
-	return errors.Join(g.ErrorInvalidRouting, fmt.Errorf("cannot route [%s] from the config node [%s]", name, r.Name()))
-}
-
-// ProceedOnFirstRoute proceeds with the first route available
-func (r *configNode) ProceedOnFirstRoute(mex f.Message) error {
-	return errors.Join(g.ErrorInvalidRouting, fmt.Errorf("cannot route from the config node [%s]", r.Name()))
-}
