+++ pkg/builders/address_book.go
@@ -1,14 +1,14 @@
 package builders
 
 import (
-	"github.com/morphy76/lang-actor/internal/routing"
-	r "github.com/morphy76/lang-actor/pkg/routing"
+	r "github.com/morphy76/lang-actor/internal/routing"
+	"github.com/morphy76/lang-actor/pkg/routing"
 )
 
 // NewAddressBook creates a new actor catalog.
 //
 // Returns:
-//   - (AddressBook): The created actor catalog.
-func NewAddressBook() r.AddressBook {
-	return routing.NewAddressBook()
+//   - (routing.AddressBook): The created AddressBook instance.
+func NewAddressBook() routing.AddressBook {
+	return r.NewAddressBook()
 }
